+++ Sources/MapboxNavigationCore/Navigator/EHorizon/RoadObjectKind.swift
@@ -6,7 +6,7 @@ extension RoadObject {
     /// Type of the road object.
     public enum Kind: Equatable, @unchecked Sendable {
         /// An alert providing information about incidents on a route. Incidents can include *congestion*,
-        /// *massTransit*, and more (see `Kind` for the full list of incident types).
+        /// *massTransit*, and more (see `Incident.Kind` for the full list of incident types).
         case incident(Incident?)
 
         /// An alert describing a point along the route where a toll may be collected. Note that this does not describe
@@ -27,23 +27,23 @@ extension RoadObject {
         /// route where a driver can choose to pull off to access a rest stop. See `MapboxDirections.StopType`.
         case serviceArea(RestStop?)
 
-        /// An alert about a segment of a route that includes a restriction. Restricted roads can include  private
+        /// An alert about a segment of a route that includes a restriction. Restricted roads can include private
         /// access roads or gated areas that can be accessed but are not open to vehicles passing through.
         case restrictedArea
 
         /// An alert about a segment of a route that includes a bridge.
         case bridge
 
-        /// An alert about an railroad crossing at grade, also known as a level crossing.
+        /// An alert about a railroad crossing at grade, also known as a level crossing.
         case railroadCrossing
 
-        /// Reserved for future use.
+        /// A road alert that was added by the user via ``RoadObjectStore/addUserDefinedRoadObject(_:)``,
         case userDefined
 
-        /// Japan-specific Interchange info, refers to an expressway entrance and exit, e.g.  Wangannarashino IC.
+        /// Japan-specific interchange info, refers to an expressway entrance and exit, e.g.  Wangannarashino IC.
         case ic(Interchange?)
 
-        /// Japan-specific Junction info, refers to a place where multiple expressways meet, e.g. Ariake JCT.
+        /// Japan-specific junction info, refers to a place where multiple expressways meet, e.g. Ariake JCT.
         case jct(Junction?)
 
         /// Undefined.